+++ payments/RELEASES.md
@@ -48,3 +48,13 @@ Based on:
 - [typescript v2.0.3] payments
 ### Releases
 - [NPM v2.0.3] https://www.npmjs.com/package/@wingspan/payments/v/2.0.3 - payments
+
+## 2023-12-11 01:29:14
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/646bee8476ac6400f0ed8faa
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.4] payments
+### Releases
+- [NPM v2.0.4] https://www.npmjs.com/package/@wingspan/payments/v/2.0.4 - payments